+++ lib/binance/spot/trade.rb
@@ -9,6 +9,7 @@ module Binance
     # - account information
     # - my trades
     # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints
+    # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints
     module Trade
       # TestNew Order
       #
@@ -29,6 +30,7 @@ module Binance
       # @option kwargs [Float] :icebergeQty
       # @option kwargs [String] :newOrderRespType Set the response JSON. ACK, RESULT, or FULL.
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#test-new-order-trade
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#test-new-order-trade
       def new_order_test(symbol:, side:, type:, **kwargs)
         Binance::Utils::Validation.require_param('symbol', symbol)
@@ -57,10 +59,18 @@ module Binance
       # @option kwargs [Float] :quoteOrderQty
       # @option kwargs [Float] :price
       # @option kwargs [String] :newClientOrderId
+      # @option kwargs [Integer] :strategyId
+      # @option kwargs [Integer] :strategyType
       # @option kwargs [Float] :stopPrice
+      # @option kwargs [Integer] :trailingDelta
       # @option kwargs [Float] :icebergeQty
       # @option kwargs [String] :newOrderRespType Set the response JSON. ACK, RESULT, or FULL.
+      # @option kwargs [String] :selfTradePreventionMode
+      # @option kwargs [String] :pegPriceType PRIMARY_PEG or MARKET_PEG.
+      # @option kwargs [Integer] :pegOffsetValue
+      # @option kwargs [String] :pegOffsetType
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#new-order-trade
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#new-order-trade
       def new_order(symbol:, side:, type:, **kwargs)
         Binance::Utils::Validation.require_param('symbol', symbol)
@@ -99,6 +109,7 @@ module Binance
       # @param kwargs [Hash]
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#cancel-all-open-orders-on-a-symbol-trade
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#cancel-all-open-orders-on-a-symbol-trade
       def cancel_open_orders(symbol:, **kwargs)
         Binance::Utils::Validation.require_param('symbol', symbol)
 
@@ -115,6 +126,7 @@ module Binance
       # @option kwargs [String] :origClientOrderId
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-order-user_data
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-order-user_data
       def get_order(symbol:, **kwargs)
         Binance::Utils::Validation.require_param('symbol', symbol)
 
@@ -128,6 +140,7 @@ module Binance
       # @param kwargs [Hash]
       # @option kwargs [String] :symbol the symbol
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#current-open-orders-user_data
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#current-open-orders-user_data
       def open_orders(**kwargs)
         @session.sign_request(:get, '/api/v3/openOrders', params: kwargs)
@@ -146,6 +159,7 @@ module Binance
       # @option kwargs [String] :endTime
       # @option kwargs [String] :limit Default 500; max 1000.
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#all-orders-user_data
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#all-orders-user_data
       def all_orders(symbol:, **kwargs)
         Binance::Utils::Validation.require_param('symbol', symbol)
@@ -185,6 +199,7 @@ module Binance
       # @option kwargs [String] :newOrderRespType
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#new-order-list---oco-trade
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#new-order-list---oco-trade
       def new_oco_order(symbol:, side:, quantity:, aboveType:, belowType:, **kwargs)
         Binance::Utils::Validation.require_param('symbol', symbol)
         Binance::Utils::Validation.require_param('side', side)
@@ -212,6 +227,7 @@ module Binance
       # @option kwargs [String] :newClientOrderId
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#cancel-order-list-trade
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#cancel-order-list-trade
       def cancel_order_list(symbol:, **kwargs)
         Binance::Utils::Validation.require_param('symbol', symbol)
 
@@ -228,6 +244,7 @@ module Binance
       # @option kwargs [Integer] :orderListId
       # @option kwargs [String] :orgClientOrderId
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-order-list-user_data
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-order-list-user_data
       def order_list(**kwargs)
         @session.sign_request(:get, '/api/v3/orderList', params: kwargs)
@@ -246,6 +263,7 @@ module Binance
       # @option kwargs [String] :limit Default 500; max 1000.
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-all-order-lists-user_data
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-all-order-lists-user_data
       def all_order_list(**kwargs)
         @session.sign_request(:get, '/api/v3/allOrderList', params: kwargs)
       end
@@ -257,6 +275,7 @@ module Binance
       # @param kwargs [Hash]
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-open-order-lists-user_data
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-open-order-lists-user_data
       def open_order_list(**kwargs)
         @session.sign_request(:get, '/api/v3/openOrderList', params: kwargs)
       end
@@ -267,7 +286,7 @@ module Binance
       #
       # @param kwargs [Hash]
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
-      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/account-endpoints#account-information-user_data
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#account-information-user_data
       def account(**kwargs)
         @session.sign_request(:get, '/api/v3/account', params: kwargs)
       end
@@ -284,7 +303,7 @@ module Binance
       # @option kwargs [Integer] :fromId TradeId to fetch from. Default gets most recent trades.
       # @option kwargs [Integer] :limit Default 500; max 1000.
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
-      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/account-endpoints#account-trade-list-user_data
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#account-trade-list-user_data
       def my_trades(symbol:, **kwargs)
         @session.sign_request(:get, '/api/v3/myTrades', params: kwargs.merge(symbol: symbol))
       end
@@ -295,7 +314,7 @@ module Binance
       #
       # @param kwargs [Hash]
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
-      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/account-endpoints#query-unfilled-order-count-user_data
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-unfilled-order-count-user_data
       def get_order_rate_limit(**kwargs)
         @session.sign_request(:get, '/api/v3/rateLimit/order', params: kwargs)
       end
@@ -311,7 +330,7 @@ module Binance
       # @param kwargs [Integer] :fromPreventedMatchId
       # @param kwargs [Integer] :limit Default: 500; Max: 1000
       # @param kwargs [Integer] :recvWindow The value cannot be greater than 60000
-      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/account-endpoints#query-prevented-matches-user_data
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-prevented-matches-user_data
       def my_prevented_matches(symbol:, **kwargs)
         Binance::Utils::Validation.require_param('symbol', symbol)
 
@@ -329,7 +348,7 @@ module Binance
       # @param kwargs [Integer] :fromAllocationId
       # @param kwargs [Integer] :limit Default 500;Max 1000
       # @param kwargs [Integer] :orderId
-      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/account-endpoints#query-allocations-user_data
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-allocations-user_data
       def my_allocations(symbol:, **kwargs)
         Binance::Utils::Validation.require_param('symbol', symbol)
 
@@ -341,11 +360,43 @@ module Binance
       # GET /api/v3/account/commission
       #
       # @param symbol [String]
-      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/account-endpoints#query-commission-rates-user_data
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#query-commission-rates-user_data
       def commission_rate(symbol:)
         Binance::Utils::Validation.require_param('symbol', symbol)
 
         @session.sign_request(:get, '/api/v3/account/commission', params: { symbol: symbol })
+      end
+
+      # Query Order Amendments (USER_DATA)
+      #
+      # GET /api/v3/order/amendments
+      #
+      # @param symbol [String]
+      # @param orderId [Integer]
+      # @param kwargs [Hash]
+      # @option kwargs [Integer] :fromExecutionId
+      # @option kwargs [Integer] :limit
+      # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/account-endpoints#query-order-amendments-user_data
+      def order_amendments(symbol:, orderId:, **kwargs)
+        Binance::Utils::Validation.require_param('symbol', symbol)
+        Binance::Utils::Validation.require_param('orderId', orderId)
+
+        @session.sign_request(:get, '/api/v3/order/amendments', params: kwargs.merge(symbol: symbol, orderId: orderId))
+      end
+
+      # Query relevant filters (USER_DATA)
+      #
+      # GET /api/v3/myFilters
+      #
+      # @param symbol [String]
+      # @param kwargs [Hash]
+      # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/account-endpoints#query-relevant-filters-user_data
+      def my_filters(symbol:, **kwargs)
+        Binance::Utils::Validation.require_param('symbol', symbol)
+
+        @session.sign_request(:get, '/api/v3/myFilters', params: kwargs.merge(symbol: symbol))
       end
 
       # Cancel an Existing Order and Send a New Order (TRADE)
@@ -375,6 +426,9 @@ module Binance
       # @option kwargs [String] :cancelRestrictions ONLY_NEW - Cancel will succeed if the order status is NEW. ONLY_PARTIALLY_FILLED - Cancel will succeed if order status is PARTIALLY_FILLED
       # @option kwargs [String] :orderRateLimitExceededMode DO_NOTHING (default)- will only attempt to cancel the order if account has not exceeded the unfilled order rate limit
       #                          CANCEL_ONLY - will always cancel the order
+      # @option kwargs [String] :pegPriceType PRIMARY_PEG or MARKET_PEG.
+      # @option kwargs [Integer] :pegOffsetValue
+      # @option kwargs [String] :pegOffsetType
       # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
       # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#cancel-an-existing-order-and-send-a-new-order-trade
       def cancel_replace(symbol:, side:, type:, cancelReplaceMode:, **kwargs)
@@ -390,6 +444,350 @@ module Binance
                                 type: type,
                                 cancelReplaceMode: cancelReplaceMode
                               ))
+      end
+
+      # Order Amend Keep Priority (TRADE)
+      #
+      # PUT /api/v3/order/amend/keepPriority
+      #
+      # @param symbol [String]
+      # @param newQty [Float]
+      # @param kwargs [Hash]
+      # @option kwargs [Integer] :orderId
+      # @option kwargs [String] :origClientOrderId
+      # @option kwargs [String] :newClientOrderId
+      # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#order-amend-keep-priority-trade
+      def amend_keep_priority(symbol:, newQty:, **kwargs)
+        Binance::Utils::Validation.require_param('symbol', symbol)
+        Binance::Utils::Validation.require_param('newQty', newQty)
+
+        @session.sign_request(:put, '/api/v3/order/amend/keepPriority',
+                              params: kwargs.merge(
+                                symbol: symbol,
+                                newQty: newQty
+                              ))
+      end
+
+      # New Order list - OCO (TRADE)
+      #
+      # POST /api/v3/orderList/oco
+      #
+      # @param symbol [String]
+      # @param side [String]
+      # @param quantity [Float]
+      # @param aboveType [String]
+      # @param belowType [String]
+      # @param kwargs [Hash]
+      # @option kwargs [String] :listClientOrderId
+      # @option kwargs [String] :aboveClientOrderId
+      # @option kwargs [Integer] :aboveIcebergQty
+      # @option kwargs [Float] :abovePrice
+      # @option kwargs [Float] :aboveStopPrice
+      # @option kwargs [Integer] :aboveTrailingDelta
+      # @option kwargs [Integer] :aboveTimeInForce
+      # @option kwargs [Integer] :aboveStrategyId
+      # @option kwargs [Integer] :aboveStrategyType
+      # @option kwargs [Integer] :abovePegPriceType
+      # @option kwargs [Integer] :abovePegOffsetType
+      # @option kwargs [Integer] :abovePegOffsetValue
+      # @option kwargs [String] :belowClientOrderId
+      # @option kwargs [Integer] :belowIcebergQty
+      # @option kwargs [Float] :belowPrice
+      # @option kwargs [Float] :belowStopPrice
+      # @option kwargs [Integer] :belowTrailingDelta
+      # @option kwargs [Integer] :belowTimeInForce
+      # @option kwargs [Integer] :belowStrategyId
+      # @option kwargs [Integer] :belowStrategyType
+      # @option kwargs [Integer] :belowPegPriceType
+      # @option kwargs [Integer] :belowPegOffsetType
+      # @option kwargs [Integer] :belowPegOffsetValue
+      # @option kwargs [Integer] :selfTradePreventionMode
+      # @option kwargs [String] :newOrderRespType
+      # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#order-amend-keep-priority-trade
+      def new_oco_order_list(symbol:, side:, quantity:, aboveType:, belowType:, **kwargs)
+        Binance::Utils::Validation.require_param('symbol', symbol)
+        Binance::Utils::Validation.require_param('side', side)
+        Binance::Utils::Validation.require_param('quantity', quantity)
+        Binance::Utils::Validation.require_param('aboveType', aboveType)
+        Binance::Utils::Validation.require_param('belowType', belowType)
+
+        @session.sign_request(:post, '/api/v3/orderList/oco',
+                              params: kwargs.merge(
+                                symbol: symbol,
+                                side: side,
+                                quantity: quantity,
+                                aboveType: aboveType,
+                                belowType: belowType
+                              ))
+      end
+
+      # New Order list - OTO (TRADE)
+      #
+      # POST /api/v3/orderList/oto
+      #
+      # @param symbol [String]
+      # @param workingType [String] Supported values: LIMIT, LIMIT_MAKER
+      # @param workingSide [String]
+      # @param workingPrice [Float]
+      # @param workingQuantity [Float]
+      # @param pendingType [String]
+      # @param pendingSide [String]
+      # @param pendingQuantity [Float]
+      # @param kwargs [Hash]
+      # @option kwargs [String] :listClientOrderId
+      # @option kwargs [String] :newOrderRespType
+      # @option kwargs [String] :selfTradePreventionMode
+      # @option kwargs [String] :workingClientOrderId
+      # @option kwargs [Float] :workingIcebergQty
+      # @option kwargs [String] :workingTimeInForce
+      # @option kwargs [Integer] :workingStrategyId
+      # @option kwargs [Integer] :workingStrategyType
+      # @option kwargs [String] :workingPegPriceType
+      # @option kwargs [String] :workingPegOffsetType
+      # @option kwargs [Integer] :workingPegOffsetValue
+      # @option kwargs [String] :pendingClientOrderId
+      # @option kwargs [Float] :pendingPrice
+      # @option kwargs [Float] :pendingStopPrice
+      # @option kwargs [Float] :pendingTrailingDelta
+      # @option kwargs [Float] :pendingIcebergQty
+      # @option kwargs [String] :pendingTimeInForce
+      # @option kwargs [Integer] :pendingStrategyId
+      # @option kwargs [Integer] :pendingStrategyType
+      # @option kwargs [String] :pendingPegPriceType
+      # @option kwargs [String] :pendingPegOffsetType
+      # @option kwargs [Integer] :pendingPegOffsetValue
+      # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#new-order-list---oto-trade
+      def new_oto_order_list(
+        symbol:,
+        workingType:,
+        workingSide:,
+        workingPrice:,
+        workingQuantity:,
+        pendingType:,
+        pendingSide:,
+        pendingQuantity:,
+        **kwargs
+      )
+        params = {
+          symbol: symbol,
+          workingType: workingType,
+          workingSide: workingSide,
+          workingPrice: workingPrice,
+          workingQuantity: workingQuantity,
+          pendingType: pendingType,
+          pendingSide: pendingSide,
+          pendingQuantity: pendingQuantity
+        }
+
+        params.each do |name, value|
+          Binance::Utils::Validation.require_param(name.to_s, value)
+        end
+
+        @session.sign_request(:post, '/api/v3/orderList/oto',
+                              params: kwargs.merge(params))
+      end
+
+      # New Order list - OTOCO (TRADE)
+      #
+      # POST /api/v3/orderList/otoco
+      #
+      # @param symbol [String]
+      # @param workingType [String] Supported values: LIMIT, LIMIT_MAKER
+      # @param workingSide [String]
+      # @param workingPrice [Float]
+      # @param workingQuantity [Float]
+      # @param pendingSide [String]
+      # @param pendingQuantity [Float]
+      # @param pendingAboveType [String]
+      # @param kwargs [Hash]
+      # @option kwargs [String] :listClientOrderId
+      # @option kwargs [String] :newOrderRespType
+      # @option kwargs [String] :selfTradePreventionMode
+      # @option kwargs [String] :workingClientOrderId
+      # @option kwargs [Float] :workingIcebergQty
+      # @option kwargs [String] :workingTimeInForce
+      # @option kwargs [Integer] :workingStrategyId
+      # @option kwargs [Integer] :workingStrategyType
+      # @option kwargs [String] :workingPegPriceType
+      # @option kwargs [String] :workingPegOffsetType
+      # @option kwargs [Integer] :workingPegOffsetValue
+      # @option kwargs [String] :pendingAboveClientOrderId
+      # @option kwargs [Float] :pendingAbovePrice
+      # @option kwargs [Float] :pendingAboveStopPrice
+      # @option kwargs [Float] :pendingAboveTrailingDelta
+      # @option kwargs [Float] :pendingAboveIcebergQty
+      # @option kwargs [String] :pendingAboveTimeInForce
+      # @option kwargs [Integer] :pendingAboveStrategyId
+      # @option kwargs [Integer] :pendingAboveStrategyType
+      # @option kwargs [String] :pendingAbovePegPriceType
+      # @option kwargs [String] :pendingAbovePegOffsetType
+      # @option kwargs [Integer] :pendingAbovePegOffsetValue
+      # @option kwargs [String] :pendingBelowType
+      # @option kwargs [String] :pendingBelowClientOrderId
+      # @option kwargs [Float] :pendingBelowPrice
+      # @option kwargs [Float] :pendingBelowStopPrice
+      # @option kwargs [Float] :pendingBelowTrailingDelta
+      # @option kwargs [Float] :pendingBelowIcebergQty
+      # @option kwargs [String] :pendingBelowTimeInForce
+      # @option kwargs [Integer] :pendingBelowStrategyId
+      # @option kwargs [Integer] :pendingBelowStrategyType
+      # @option kwargs [String] :pendingBelowPegPriceType
+      # @option kwargs [String] :pendingBelowPegOffsetType
+      # @option kwargs [Integer] :pendingBelowPegOffsetValue
+      # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#new-order-list---otoco-trade
+      def new_otoco_order_list(symbol:, workingType:, workingSide:, workingPrice:, workingQuantity:, pendingSide:, pendingQuantity:, pendingAboveType:, **kwargs)
+        params = {
+          symbol: symbol,
+          workingType: workingType,
+          workingSide: workingSide,
+          workingPrice: workingPrice,
+          workingQuantity: workingQuantity,
+          pendingSide: pendingSide,
+          pendingQuantity: pendingQuantity,
+          pendingAboveType: pendingAboveType
+        }
+
+        params.each do |name, value|
+          Binance::Utils::Validation.require_param(name.to_s, value)
+        end
+
+        @session.sign_request(:post, '/api/v3/orderList/otoco',
+                              params: kwargs.merge(params))
+      end
+
+      # New Order List - OPO (TRADE)
+      #
+      # POST /api/v3/orderList/opo
+      #
+      # @param symbol [String]
+      # @param workingType [String] Supported values: LIMIT, LIMIT_MAKER
+      # @param workingSide [String]
+      # @param workingPrice [Float]
+      # @param workingQuantity [Float]
+      # @param pendingType [String]
+      # @param pendingSide [String]
+      # @param kwargs [Hash]
+      # @option kwargs [String] :listClientOrderId
+      # @option kwargs [String] :newOrderRespType
+      # @option kwargs [String] :selfTradePreventionMode
+      # @option kwargs [String] :workingClientOrderId
+      # @option kwargs [Float] :workingIcebergQty
+      # @option kwargs [String] :workingTimeInForce
+      # @option kwargs [Integer] :workingStrategyId
+      # @option kwargs [Integer] :workingStrategyType
+      # @option kwargs [String] :workingPegPriceType
+      # @option kwargs [String] :workingPegOffsetType
+      # @option kwargs [Integer] :workingPegOffsetValue
+      # @option kwargs [String] :pendingClientOrderId
+      # @option kwargs [Float] :pendingPrice
+      # @option kwargs [Float] :pendingStopPrice
+      # @option kwargs [Float] :pendingTrailingDelta
+      # @option kwargs [Float] :pendingIcebergQty
+      # @option kwargs [String] :pendingTimeInForce
+      # @option kwargs [Integer] :pendingStrategyId
+      # @option kwargs [Integer] :pendingStrategyType
+      # @option kwargs [String] :pendingPegPriceType
+      # @option kwargs [String] :pendingPegOffsetType
+      # @option kwargs [Integer] :pendingPegOffsetValue
+      # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#new-order-list---opo-trade
+      def new_opo_order_list(symbol:, workingType:, workingSide:, workingPrice:, workingQuantity:, pendingType:, pendingSide:, **kwargs)
+        params = {
+          symbol: symbol,
+          workingType: workingType,
+          workingSide: workingSide,
+          workingPrice: workingPrice,
+          workingQuantity: workingQuantity,
+          pendingType: pendingType,
+          pendingSide: pendingSide
+        }
+
+        params.each do |name, value|
+          Binance::Utils::Validation.require_param(name.to_s, value)
+        end
+
+        @session.sign_request(:post, '/api/v3/orderList/opo',
+                              params: kwargs.merge(params))
+      end
+
+      # New Order List - OPOCO (TRADE)
+      #
+      # POST /api/v3/orderList/opoco
+      #
+      # @param symbol [String]
+      # @param workingType [String] Supported values: LIMIT, LIMIT_MAKER
+      # @param workingSide [String]
+      # @param workingPrice [Float]
+      # @param workingQuantity [Float]
+      # @param pendingSide [String]
+      # @param pendingAboveType [String]
+      # @param kwargs [Hash]
+      # @option kwargs [String] :listClientOrderId
+      # @option kwargs [String] :newOrderRespType
+      # @option kwargs [String] :selfTradePreventionMode
+      # @option kwargs [String] :workingClientOrderId
+      # @option kwargs [Float] :workingIcebergQty
+      # @option kwargs [String] :workingTimeInForce
+      # @option kwargs [Integer] :workingStrategyId
+      # @option kwargs [Integer] :workingStrategyType
+      # @option kwargs [String] :workingPegPriceType
+      # @option kwargs [String] :workingPegOffsetType
+      # @option kwargs [Integer] :workingPegOffsetValue
+      # @option kwargs [String] :pendingAboveClientOrderId
+      # @option kwargs [Float] :pendingAbovePrice
+      # @option kwargs [Float] :pendingAboveStopPrice
+      # @option kwargs [Float] :pendingAboveTrailingDelta
+      # @option kwargs [Float] :pendingAboveIcebergQty
+      # @option kwargs [String] :pendingAboveTimeInForce
+      # @option kwargs [Integer] :pendingAboveStrategyId
+      # @option kwargs [Integer] :pendingAboveStrategyType
+      # @option kwargs [String] :pendingAbovePegPriceType
+      # @option kwargs [String] :pendingAbovePegOffsetType
+      # @option kwargs [Integer] :pendingAbovePegOffsetValue
+      # @option kwargs [String] :pendingBelowType
+      # @option kwargs [String] :pendingBelowClientOrderId
+      # @option kwargs [Float] :pendingBelowPrice
+      # @option kwargs [Float] :pendingBelowStopPrice
+      # @option kwargs [Float] :pendingBelowTrailingDelta
+      # @option kwargs [Float] :pendingBelowIcebergQty
+      # @option kwargs [String] :pendingBelowTimeInForce
+      # @option kwargs [Integer] :pendingBelowStrategyId
+      # @option kwargs [Integer] :pendingBelowStrategyType
+      # @option kwargs [String] :pendingBelowPegPriceType
+      # @option kwargs [String] :pendingBelowPegOffsetType
+      # @option kwargs [Integer] :pendingBelowPegOffsetValue
+      # @option kwargs [Integer] :recvWindow The value cannot be greater than 60000
+      # @see https://developers.binance.com/docs/binance-spot-api-docs/rest-api/trading-endpoints#new-order-list---opoco-trade
+      def new_opoco_order_list(
+        symbol:,
+        workingType:,
+        workingSide:,
+        workingPrice:,
+        workingQuantity:,
+        pendingSide:,
+        pendingAboveType:,
+        **kwargs
+      )
+        params = {
+          symbol: symbol,
+          workingType: workingType,
+          workingSide: workingSide,
+          workingPrice: workingPrice,
+          workingQuantity: workingQuantity,
+          pendingSide: pendingSide,
+          pendingAboveType: pendingAboveType
+        }
+
+        params.each do |name, value|
+          Binance::Utils::Validation.require_param(name.to_s, value)
+        end
+
+        @session.sign_request(:post, '/api/v3/orderList/opoco',
+                              params: kwargs.merge(params))
       end
     end
   end